Move the `util-client` module under the `@bayou` namespace.
